Simplify DragSource connection
