+++ src/TableColumn.js
@@ -3,7 +3,22 @@ import ReactDOM from 'react-dom';
 import Const from './Const';
 import { DragSource } from 'react-dnd';
 import RowTypes from './RowTypes';
-import rowSource from './RowSource';
+
+const rowSource = {
+  beginDrag(props) {
+    return {
+      rIndex: props.rIndex
+    };
+  },
+
+  endDrag(props, monitor) {
+    if (props.onDroppedRow) {
+      const dragIndex = props.rIndex;
+      const dropIndex = monitor.getItem().rIndex;
+      props.onDroppedRow(dragIndex, dropIndex);
+    }
+  }
+};
 
 @DragSource(RowTypes.ROW, rowSource, (connect, monitor) => ({
   connectDragSource: connect.dragSource()
